requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
pydantic==1.10.7
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
cryptography==40.0.1
mysqlclient==2.1.1
pymysql==1.0.3
websockets==11.0.1